requirements.txt
telethon==1.36.0
python-telegram-bot==21.5
gspread==6.1.2
google-auth==2.34.0
python-dotenv==1.0.1
